commits: don't add extra blanks if summary is empty (Bug 1453669).
Summary:
We now only include blanks for sections of the commit message which are
not empty. This results in only a single blank line between the commit
message title and the differential revision link when the summary is
empty.

Test Plan: `invoke test` passes, new test added which fails before fix.

Reviewers: mars

Reviewed By: mars

Bug #: 1453669

Differential Revision: https://phabricator.services.mozilla.com/D1031