repos: configure landing ldap group per repo (Bug 1452687).
Summary:
The configuration for each repo now includes an `AccessGroup`,
specifying the ldap group required to land to the repository. Each
repository has been updated with an appropriate group, and the
`SCMLevelInsufficient` blocker now checks for membership in the
configured group.

The `SCMLevelInsufficient` is now checked after `InvalidRepository` and
will not cause a blocker if configuration doesn't exist for the current
repository. This should be okay since the `InvalidRepository` blocker
will prevent landing in these cases.

Test Plan: `invoke test` passes, tests updated.

Reviewers: mars

Reviewed By: mars

Bug #: 1452687

Differential Revision: https://phabricator.services.mozilla.com/D1032