reviewers: temporarily fix  project reviewers (Bug 1456985).
Summary:
Lando is breaking when a project is set as a reviewer, as it expects
to get information for each reviewer from 'user.search'. This quick
fix puts us back to defaulting to a username in this case, this time
using '<unknown>' instead of empty string ''.

Test Plan: `invoke test` passes.

Reviewers: ckolos, ckolos_admin

Reviewed By: ckolos_admin

Subscribers: ckolos_admin

Bug #: 1456985

Differential Revision: https://phabricator.services.mozilla.com/D1040