landings: don't race on landing submission (Bug 1441269).
Summary:
Before this change it would be possible for two callers to the landings
endpoint to race on submission to transplant. If both requests came in
at the same time they might both see no submitted landings, and then
submit a landing themselves.

We now lock the landings table to other landing submissions before
checking for a previous submission and submitting a landing. Other
normal reads of the landings table should be allowed while this lock is
held, only reads that explicitly request the same lock mode and writes
will have to wait on the lock.

Unfortunately this means all landing requests to transplant will be made
serially, as incoming requests to /landings will wait on the lock. Lando
will also have to wait on this lock to write transplant status updates
to the db. I anticipate the performance impact to be quite small, as
hopefully the critical section is executed quickly.

In the future if the performance impact becomes too great we could
restructure the critical section to only write a flag row to the
database. This would mean we could make the request to transplant
outside the critical section. I chose to go with the larger critical
section here because it was much simpler; there is no need to have a
mechanism for cleaning up old flag rows that would block all future
landings.

The filename used for patches uploaded to S3 was also changed to remove
the landing id. This meant that the patch upload could take place
before the critical section and creating an aborted landing isn't
required.

Depends on D1047.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1441269

Differential Revision: https://phabricator.services.mozilla.com/D1049