Allow landing the `test-repo` if in `localdev` environment.
Summary:
`localdev` environment is used by default in Suite.  Also Lando API
requires it to be set for mocking the auth0 user. Current configuration
does not allow to land any repository if in `localdev` environment.

This patch is adding the `test-repo` as an allowed repository in
`localdev` environment and sets the reqiured commit access to Level 1.

Test Plan: Check if `test-repo` is landable in `localdev` environment.

Reviewers: smacleod

Reviewed By: smacleod

Subscribers: smacleod

Bug #: 1470398

Differential Revision: https://phabricator.services.mozilla.com/D1764