landings: block older diffs instead of warning (Bug 1456859).
Summary:
It's rare that you would ever want to land an older diff and having
support for doing so complicates things quite a bit. This change takes
the first step in removing older diff landing support by preventing it
with a blocker. Future changes may now remove the functionality /
simplify things as appropriate.

If a user really does need to land an older version of a diff it is
still possible by downloading the version and creating a new diff with
it.

Test Plan: `invoke test` passes, tests updated and obsolete tests removed.

Reviewers: glob

Reviewed By: glob

Bug #: 1456859

Differential Revision: https://phabricator.services.mozilla.com/D1811