stacks: add function to build a graph of a stack (Bug 1473127).
Summary:
`build_stack_graph` takes a revision PHID and builds a graph
representing the entire stack it is part of. Any revision that is part
of the stack should result in an equal graph.

This function is expensive to call as it can make many requests to
Phabricator in a single call. In the future we might want to consider
caching things where possible.

Test Plan: `invoke test` passes, new tests.

Reviewers: glob

Reviewed By: glob

Bug #: 1473127

Differential Revision: https://phabricator.services.mozilla.com/D2087