stacks: add stacks endpoint (Bug 1473127).
Summary:
This introduces the basic stack querying endpoint. There are three main
components at this time: The revisions that are part of the stack, the
edges between each revision forming the stack graph, and a list of
"landable paths".

A landable path is a series of revisions forming a line through the
stack DAG. Each of these landable paths meets a set of criteria required
for landing (such as not containing revisions with more than one open
parent).

Data on the status of each revision and specific repositoriy data for
each revision has been left out as it is not needed at this time.
Introducing more fields to the endpoint for this data will come in a
future change.

Depends on D2088.

Test Plan:
`invoke test` passes, new tests added. Interactively requested various
revision stacks from phabricator-dev.

Reviewers: glob

Reviewed By: glob

Bug #: 1473127

Differential Revision: https://phabricator.services.mozilla.com/D2089