transplants: add dryrun endpoint (Bug 1473127).
Summary:
This change provides a new transplant dryrun endpoint, laying out the
initial structure. Some features are missing, such as the proper
generation of a confirmation token and further blockers / warnings, but
enough is provided to allow UI to be developed. Future changes will
iterate on this endpoint and integrate it with the new Transplant
submission endpoint.

Since this new dryrun deals with many revisions at once, it aggregates
warning by type while still allowing details particular to a single
revision. The idea behind this being that a UI may show and allow
acknowledging the warnings of a single type all at once, rather than
individually for every revision with that warning. It is still possible
to drill down into specific details if needed.

Depends on D3270.

Test Plan: `invoke test` passes, tests added. Observed results in series UI.

Reviewers: glob

Reviewed By: glob

Bug #: 1473127

Differential Revision: https://phabricator.services.mozilla.com/D3269