python: update requirements (Bug 1475614).
Summary:
Update the python requirements to their latest versions. 'yapf' was not
updated as part of this sweep since it wanted to rewrite most of our
code (and we're going to switch to black soon).

Since the dev and common requirements files were including a bunch of
the same dependencies and most of the dev dependencies aren't all that
heavy, I've combined them into a single base.txt file. The production
file is still kept as we don't want something large like uwsgi in dev.

Test Plan:
`invoke test` passes after a `docker-compose build`. Ran the devserver
and tested a couple endpoints to ensure that was working as well.

Reviewers: glob

Reviewed By: glob

Subscribers: mcote, ulfr, glob, gps, smacleod, ckolos, hwine

Tags: #secure-revision, #bmo-conduit-security

Bug #: 1475614

Differential Revision: https://phabricator.services.mozilla.com/D2289