stacks: expose why revisions are blocked (Bug 1473127).
Summary:
This change exposes blocker data through the stack API for revisions
which don't appear in a landable path. Each revision that is not
landable has a reason for being blocked, this is sent in the
"blocked_reason" field for each revision.

The landable sugraphs calculation has been updated to track this blocker
information and now returns it. An optional `other_checks` parameter has
been added to the function to support future non graph based reasons
for a revision being blocked. The argument should be an iterable of
callables which will be run against revisions to determine if they
should be blocked. This design was chosen to make testing both the
landable sugraph calculation and future blocker checks easier.

The downside of the current design means any blocker checks that depend
upon the structure of the graph must be part of the main algorithm. I
couldn't think of any other graph structure checks we'd ever want and
this helped simplify things considerably.

Depends on D3267.

Test Plan: `invoke test` passes.

Reviewers: glob

Reviewed By: glob

Bug #: 1473127

Differential Revision: https://phabricator.services.mozilla.com/D3268