db: migrate from Landing to Transplant (Bug 1473127).
Summary:
This change migrates the codebase from the Landing model to a new
Transplant model which can support landings involving multiple
revisions. As part of the migration all previous Landings will be
represented with a new equivalent Transplant entry. We consciously throw
away Landing.active_diff_id as we no longer support landing "inactive"
diffs and the history isn't important.

The migration leaves the old Landing database table and data around as a
safety net. It can be cleaned up in a future change after this has been
successfully deployed.

All uses of Landing have been replaced with equivalent Transplant usage.
This will allow the current single revision Lando to operate while the
new series support is tested / developed. Eventually when series support
is ready to completely replace the current API legacy code may be
removed.

The most interesting difference between Landing and Transplant is the
replacement of `revision_id` and `diff_id` with `revision_to_diff_id`
and `revision_order`. `revision_to_diff_id` is a JSONB field, the
contents being a JSON object mapping revision ids to the diff id used
for that revision. This allows a single Transplant row to contain
indexable, queryable information regarding the set of revisions in a
series. Particularly it is possible to query for all Transplants that
contain any revision from a set (such as a stack). The `revision_order`
JSONB field is meant to support `revision_to_diff_id`, providing the
ordering for the revisions in the Transplant. This is needed since JSONB
objects do not maintain key ordering.

Depends on D3268.

Test Plan:
`invoke test` passes. Tested the database migration with full landings
table dumps from both dev and prod lando-api.

Reviewers: glob

Reviewed By: glob

Bug #: 1473127

Differential Revision: https://phabricator.services.mozilla.com/D3270