commit_message: detect group reviewer hash tags (bug 1512726)
Summary:
Prior to this commit, lando did not properly strip group reviewer hash
tags from the final commit message. This is because these group reviewer
hash tags started with a '#'. With this commit, we properly check for
this when matching reviewers in the commit message.

Test Plan: new test added, invoke test passes

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1512726

Differential Revision: https://phabricator.services.mozilla.com/D14711