email: add configuration for smtp ssl, tls, and auth (Bug 1535759).
Summary:
This adds the other `MAIL_*` environment variables that Ops was
expecting and a ready check to ensure things are configured properly.

Test Plan:
`invoke test` passes. Ran manually to verify the readiness check and
and that I could send to the SMTP server. Also ran with an alternative
STMP server that supported auth and made sure it worked.

Reviewers: mars, imadueme

Reviewed By: mars

Bug #: 1535759

Differential Revision: https://phabricator.services.mozilla.com/D24756