notifications: don't notify when tree is closed.
Summary:
Lando was treating any message from transplant where the change hadn't
landed as something to notify about. Transplants sends messages like
this when it has encountered an intermittent failure (such as the trees
being closed). The proper field to determine an unrecoverable error is
the "error_msg" string.

This change switches Lando to gate the notification logic on the
Transplant model status, which takes the strange error conventions into
account and exposes it as failed.

Test Plan:
`invoke test` passes. Test added that fails before change as well as a
test for the happy path.

Reviewers: mars

Reviewed By: mars

Differential Revision: https://phabricator.services.mozilla.com/D25634