transplants: remove #check-in_needed when landing (Bug 1535157) (#65).
Because Check-in Needed is being migrated to Phabricator and will be
handled by Lando, we now remove the tag from revisions when landing.
This is done asynchronously after triggering the landing so that any
errors will not block a proper response. It could have been done before
triggering the landing but any transient failure talking to transplant
would have prevented the landing even though the tag was removed.

Additionally Lando now has a `PHABRICATOR_ADMIN_API_KEY` environment
variable which is used to specify an API key for privileged actions.
This key must have access to all revisions, including those marked as
secure. Care should be taken when using this API key.

Removal of the #check-in_needed tag uses the admin api key to ensure
that it is successful for all revisions and that it is clear Lando is
the bot user causing the removal.

This change was tested by updating the automated tests, and running the
removal code manually against phabricator-dev.allizom.org.