dependencies: upgrade all dependencies.
This change upgrades all of the python dependencies and moves lando api
to python 3.7. The standard python docker image is now used to greatly
simplify dependency installation and better take advantage of shared
layer caching between applications.

The largest change that had to be worked around was in connexion. It no
longer supports having handlers themselves take care of security
schemes, and forbids the use of multiple schemes on a single endpoint.
This is problematic because we allow anonymous access to some endpoints
that also accept an api key. Integrating our authentication with the new
built-in auth wrappers would be quite a lot of work and wouldn't even be
properly possible without support for multiple schemes. Because of this
all security definitions have been removed from the swagger spec. While
unfortunate, they are only informational when running in production.