app: only wrap API exceptions in production environments
The default error handlers eat the PhabricatorAPIException root cause
and traceback. This makes debugging API failures difficult and breaks
PyCharm unittest integration.

We can make the error handler optional in the same way that the Flask
default error handler is optional: in production it wraps all errors,
but in the test and debug environments exceptions are allowed to
propagate through the stack to make debugging easier.