stacks: provide a more specific error message when no repo is specified for a revision (Bug 1569721)(#75)
In `calculate_landable_subgraphs` we iterate over the repository
values attached to the revisions in our stack and check if the
repo is valid using `repo not in landable_repos`. This blocks
the revision from landing and displays a sensible error message
about an unsupported repo.

However, in the case a repo has not been associated with a
revision, the condition will still be true as the empty value is
not in the set of landable repos. This causes confusion to end
users regarding what to do to fix the problem. In this commit we
add an extra check to assert the value of `repo` is not a
False-y value. If the value is False-y, we provide a more instructive
error message and block the revision from landing. Otherwise,
we continue with the existing block logic.