models: add a Base model class for convenience
Note: this change also updates SecApprovalRequest to have an updated_at
field.

The `Base` class can be extended to create a model that has some common
attributes and fields, namely: `updated_at` and `created_at` fields, an
`id` field, a default `__tablename__` (snake-case of the class name), as
well as a default `__repr__` method.