Bug 1548114: implement autoformatting in Lando
Implements autoformatting in Lando via `hg fix`. Autoformatting runs
at landing time and changeset hashes which are replacements for original
hashes are stored on the landing job object. Commands for formatters are
defined in the Lando config for a repo while fileset patterns are
defined in a `.lando.ini` file in-tree.

Updates Mercurial to 5.5.1 and adds a few minor changes for the update.
We need to update Mercurial to a newer version to get a minor bugfix in
`hg fix`. We remove the `-r` flag from the call to `rebase --abort`
since the flags are mutually exclusive on newer versions of Mercurial.