Bug 1749754: add a landing warning to check if a patch has `WIP:` in the title
Adds a revision warning to assert a revision does not have `WIP:`
as the beginning of it's revision title. The presence of `WIP:`
indicates a work-in-progress revision and should be scrubbed before
landing by the user.

We could follow up with some code to automatically scrub the `WIP`
text from the commit message, but this is a good start for now.