Bug 1745293: amend release managers review check to check all required uplift conditions (#180)
This commit extends the existing uplift blocking check to include
checking for a submitted uplift request form somewhere on the
stack. The function name is updated to reflect the change.

We also add code to add the commit message rewriting paths to
change `r=release-managers,relman-reviewer` to `a=relman-reviewer`
to conform with existing syntax in uplit repositories.