Bug 1754801: compare enum equality with `==` consistently (#182)
Lando makes use of enums to compare allowed values,
and is inconsistent about using `==` or `is` to
make these comparisons. Convert uses of `is` enum
comparisons to `==` wholesale.

Checked by doing a global search for all `@enum`
in the repo, then searching for all instances of
each class type and ensuring comparisons are made
with `==`.