treestatus: consider "approval required" status as open for landing (Bug 1756036) (#185)
Within the new uplift workflow, repositories on Lando with
"approval required" status should be considered open for landing.
The approval is checked in Lando by asserting `release-managers`
review on the revision and the hg hook asserts `a=<reviewer>` is
present in the first line of the commit message. Thus we can
consider an "approval required" repo as landable and enabled.