approvals: only rewrite RelMan reviews as `a=` for approval required repositories (Bug 1755932)
Fixes a regression caused by #180 where all RelMan reviews are
rewritten from `r=reviewer` to `a=reviewer`. This should only
occur on an uplift repository, so we amend usages of
`approvals_for_commit_message` to only apply when this flag is
set.

Some of the secure revision tests fail after this change due to
a missing repository tagged on the revision so we add a repo to
these tests.