autoformat: change from `hg fix` to tip commit/amend strategy (Bug 1785004)
Remove the `hg fix` based autoformatting and replace with
running raw `mach lint` on the tip commit, creating a tip
autoformat commit for stacks of changes and amending the
commit for landings of a single changeset. Add a call to
`mach bootstrap` in the repo clone subsystem when autoformat
is enabled for a given repo, to enabled downloading required
packages. Since we now run `mach` from within the landing repo,
stop running `hg purge` with `--all` since that will remove
the existing objdir required for linting virtualenvs. Add the
`MOZBUILD_STATE_PATH` environment variable to the Dockerfile
to avoid a prompt during `mach bootstrap`.