uplift: fix endpoint and add support for stack uplifts (Bug 1764213) (#200)
uplift: make uplift endpoint work for a single revision (Bug 1764213)

Add support for creating a stack of uplift revisions based on an existing
Phabricator stack. Adds two API endpoints. The first returns the list of
repositories tagged with `uplift` for display as a repo in Lando. The
second creates a stack of Phabricator revisions for uplift with
`release-managers` tagged as a Phab review group.