uplift: use `packaging.version.Version` to handle milestone version parsing (Bug 1788078)
The regex used to parse the major milestone version failed
on the content of the current beta (which is now in the tree
as a test case as `MILESTONE_TEST_CONTENTS_2`). This commit
changes the parsing to instead take the last line of the file
and pass to a `Version` object, which is returned from the
parsing function.