phabricator: call `diffusion.looksoon` on landing to trigger repo sync (Bug 1455468, Bug 1788722) (#216)
Adds a call to Phabricator's `diffusion.looksoon` API to trigger
syncing from the upstream repo. This should cause revisions
to close faster, especially on uplift repositories.

As part of this patch some tests were failing due to out of date
`moto` conflicting with `responses`. Upgrading `moto` caused the
removal of several packages that are used directly in testing and
the main application. These packages are added to `requirements.in`
so they are explicit dependencies.