typing: set `scopes` type to `Iterable` (Bug 1759890)
The `scopes` type is set to `tuple[str]`, which implies to type
checkers that the type is always a single `str` element tuple.
Since we are passing `tuple`'s of different length in various
places, it is more appropriate for the type to be `Iterable`,
which also aligns with the docstring comment about `scopes`.

Use the `Iterable` from `collections.abc` as that is the
preferred way in Python 3.9+.