uplift: stop using a `landing_path` for uplift endpoint (Bug 1801926)
Change the swagger schema for the uplift endpoint to take
a revision ID instead of a landing path. Add simple stack
walking methods to `RevisionStack` and make `get_uplift_conduit_state`
return a `RevisionStack`. Use these stack walking methods in
place of the passed `landing_path` to determine the order
in which to create uplift revisions.