phabricator: remove `g.phabricator` and replace with arguments to command function (Bug 1759890) (#236)
Move the `PhabricatorClient` from `g.phabricator` to an argument
passed to the necessary API function. This makes cases where
we require Phabricator more explicit and removes the global
`g.phabricator`, which was inconsistently defined and typed.

In making this change we also notice the `get_phabricator_repo`
usage and function is unnecessary as we have already retrieved
the required `repo` information from Phabricator when the function
is called.

Also rename some variables where we convert between types to
keep each variable name consistently typed.