projects: remove `get_relman_phid` in favor of `get_release_managers` (Bug 1806455)
In a few places within Lando-API we retrieve the `#release-managers`
group information. In some cases we only require a PHID and in some
cases we need the entire group information. When we need each, we
make a call to the Conduit API to get the required info.

Consolidate these two use cases into one function that returns all
information for the `release-managers` project, eliminating the extra
calls to Conduit. Move `get_release_managers` to `projects.py` where
is makes more sense and cache it for efficiency.

Add the `release_management_project` fixture to `test_transplants.py`
tests where it is needed.