api, models: finish adding type hints, make `ConfigurationKey` a proper enum (Bug 1759890) (#261)
Finish adding type hints to all modules in `api/` and `models/`.

Type hints for `put` are updated from `int` to `str` to reflect
the correct state in `swagger.yml`.

Add a type alias for configuration values that is an alias to
`bool | int | str`. Also raise at the end of the function to
avoid an accidental `None` return in some edge cases.

Add `sqlalchemy.orm.Query` as the return type for model queries.