stacks: implement `RevisionStack` with `networkx` (Bug 1811469)
Re-implement `RevisionStack` with `networkx`. Create a `graph`
attribute that holds the edge/node data and switch calls to
`child` and `parent` mappings to instead use `predecessor`
and `successor`, bringing it more in line with the usual
terminology used within Conduit.

`base_revisions` is updated to instead iterate over the
elements in the graph with an in-degree of zero. The
in-degree is the number of directed edges that point
to the node (ie the number of predecessors).

Add a `head` argument to `iter_stack_from_base` to set
the "destination" node and change the function to use
the "simple paths" algorithm from `networkx`.