stacks: implement `RevisionStack` with `networkx` (Bug 1811469) (#266)
Re-implement RevisionStack with networkx. Implement
the class as a subclass of nx.DiGraph to use its representation.
Update child and parent mappings to instead use predecessor
and successor, bringing it more in line with the usual
terminology used within Conduit.

base_revisions is updated to instead iterate over the
elements in the graph with an in-degree of zero. The
in-degree is the number of directed edges that point
to the node (ie the number of predecessors). We also
rename the function to root_revisions to stay aligned
with the networkx graph terminology.

Add a dest argument to iter_stack_from_base to set
the "destination" node and change the function to use
the "simple paths" algorithm from networkx. Rename
the function to iter_stack_from_root to align with
networkx terminology.