transplants: separate `None` check for required variables (Bug 1759890)
Move `None` check for required variables into their own
`if` statements. This makes the exception more explicit
about what values are missing as the previous implementation
would print a `None` to the screen instead of showing
which variable could not be resolved. This also has the
effect of letting type checkers know the value is not
`Optional` after that check.