projects: clarify that `get_release_managers` returns an `Optional` (Bug 1759890)
Add an `Optional` type around `get_release_managers`, since it
returns the value of `phab.single` which is not confirmed to
exist. Add a `None`-check to locations where we assume it is
available.

While we're here also add a type hint to `project_search` as
it is missing one.