phabricator: cleanup some of the patch parsing code (#305)
While trying to determine if we had a bug in the
code that converts a Phabricator diff into a format
that is usable for uploading as a Phabricator revision
I found some areas that could be cleaned up in
`phabricator_patch.py`.

Move building the `news` and `olds` lists into the
main loop. Remove brackets around the variable names
in the for loop. Add periods to comments for some
consistency. Add docstrings to enums where appropriate.
Create an enum for diff operations and use it in the
diff parsing logic.