auth: make `scopes` a required argument in `require_auth0` decorator (Bug 1832362)
Change `scopes` from a keyword argument to a positional argument,
indicating that scopes must be passed to the decorator. The keyword
argument syntax of `scopes=` still works for positional arguments
so no other changes are required and the purpose of the argument
is still clear when reading usages of the decorator.