landing_job: move patch fetching to `LandingJob` (Bug 1836499)
The `all_patches` list is built in the `LandingWorker` as
the ordered set of patches to apply to the repo for a job.
The contents of this list is built entirely from fields in
the `LandingJob`, so move the fetching into a function
of the job. This will enable us to abstract details of
the patch generation away in future work, and also makes
the code cleaner.

The fetching function in the job is made into a generator
instead of returning a list, which is a slight optimization
as we no longer retrieve and create a `BytesIO` for patches
which would be applied after a merge conflict or other error.
Since we no longer have a list of patches to check the length
of for autoformatting, use the length of the changeset titles
list instead, as it is a more accurate number in that
context anyways.

A simple test is added for the new `get_patches` function.
While adding the test the revision creation test fixture is
moved to `conftest.py` so it can be used from the new test.
In doing so we move the `PATCH_NORMAL_1` patch to the new
file, and remove the unused `PATCH_NORMAL_2`. We rename
`PATCH_NORMAL_3` to a more useful name to reflect the removal
of the other patch variables.