typing: add types to `hgexports.py` and fix `set_patch` types (Bug 1759890) (#330)
While testing the try API endpoint I noticed my very simple
patch was failing to apply. This is because the type hint for
`Revision.set_patch` is incorrect and the raw diff returned
from Phabricator is actually a `str`, not `bytes`, and thus the
type I was using was incorrect. The timestamp is also a `str`
and not an `int`. Update the function signature to match reality,
and while we're here add type hints to all of `hgexports.py`.