api: add an API endpoint for submission of try pushes (Bug 1835959) (#328)
Adds a new POST /try endpoint to Lando that accepts try patch
data and creates a landing job to submit the patches to try.
Patches are submitted as an array of base64 encoded bytes and
the base commit for the try submission in mozilla-unified is
passed so the patch applies as it would in the developer's local
machine.

The Repo object sees several changes to allow Lando to work
with the try repo. A force_push option is added that controls
adding the -f option during hg push. A phabricator_repo
flag is also added which indicates if the repo is present on
Phabricator. At the moment this flag controls pinging Phabricator
to issue a repo update API call, which is irrelevant for try.

The try repo is added to the repo configs, with mozilla-unified
as the pull_url and try as the push_url. This means patches
will be applied to mozilla-unified in Lando and pushed out to
try, in the same fashion as the current try push model on a
developer's local machine.

A decorator is added to enforce active_scm_level_1 group
permissions with appropriate error messages.

Patches can be submitted in the hg export or git format-patch
formats. Support for git format-patch is added via a Git patch parser
similar to the HgPatchHelper.