hg: add a `PullFailureException` to retry landings after a 500 during pull (Bug 1858500) (#349)
Add a `PullFailureException` with a snippet that looks for errors
during `hg pull` and add it to the list of Lando exceptions
that can be raised from an hglib exception. Add an except
case to the `update_repo` try block that is similar to the case
in the `push` try block, to allow the job to be retried when this
exception is encountered.