hg: change `PullFailureException` to `HgmoInternalServerError` (Bug 1860067)
The error snippet for `PullFailureException` is a piece of text that
comes from Lando, not Mercurial. The correct piece of text to look
for is `abort: HTTP Error 500`, which is a more general error that
can happen when load on hg is high.

Rename `PullFailureException` to `HgmoInternalServerError` and update
the snippet to reflect the output from hg. Since this error is more
general, we also add it to the set of exceptions to consider a temporary
failure at push time.