emails: make landing failure emails less Phab-specific (Bug 1855851)
Make the Phabricator landing emails less Phabricator-specific.
Update the landing job identifier to include the context that
the job belongs to a try push with the included commit message
title. Add a check of the landing job identifier to only display
a link to a Lando-UI URL when the identifier is a Phab revision
ID.