requirements: update various requirements including PyYAML (Bug 1864469) (#356)
PyYaml <= 6.0.0 has a bug that causes build failues on later versions of
Python due to a Cython version bump. I noticed this when I forgot to source
a virtualenv and attempted to install the Lando-API requirements into a
Python 3.12 interpreter. The fix is available on PyYaml 6.0.1, so we should
get ahead of this problem and upgrade requirements now while it's low risk.