Bug 1867036: empty Git commits aren't parsed correctly
When empty Git commits are sent to Lando in `git format-patch`
format, Lando fails to parse the commit correctly. Currently
Lando will fail to find the `---` line used to separate the
commit message from the diff, and assume all lines in the
patch body are part of the commit message. The parsing will
then fail later down the line as Lando will not find a valid
diff line in the body.

Add an `else` clause to the loop which looks for the `---` line
separating the extended commit message and the diff, which when
reached indicates that no diff is present on the commit. In this
case we strip the last two lines of the gathered commit message
and return an empty diff. Add a test for parsing a patch generated
from an empty commit to `test_hgexports.py`.