uplifts: query `differential.querydiffs` to find the latest non-`commit` diff (Bug 1709608)
Extend `get_uplift_conduit_state` to also query `differential.querydiffs`
and return all known diffs associated with the selected uplift
revisions. When creating the uplift revisions, add a function to
parse through the diffs and return the latest diff that does not
have a `creationMethod` of `commit`. This works around an issue
in Phabricator where diffs are created with empty binary files
when a new diff is created after parsing a commit from the observed
version control repository.